+++ gdpc/utils.py
